Move ProcessQueue biz logic into viewmodel
