+++ Source/LibationWinForms/AvaloniaUI/Views/MainWindow/MainWindow.axaml
@@ -8,7 +8,9 @@
 		xmlns:prgid="clr-namespace:LibationWinForms.AvaloniaUI.Views.ProductsGrid"
 		xmlns:controls="clr-namespace:LibationWinForms.AvaloniaUI.Controls"
 		mc:Ignorable="d" d:DesignWidth="2000" d:DesignHeight="700"
-		x:Class="LibationWinForms.AvaloniaUI.Views.MainWindow" Title="MainWindow">
+		x:Class="LibationWinForms.AvaloniaUI.Views.MainWindow"
+		Title="MainWindow"
+		Icon="/AvaloniaUI/Assets/glass-with-glow_16.png">
 
 	<Border BorderBrush="{DynamicResource DataGridGridLinesBrush}" BorderThickness="2" Padding="15">
 		<Grid RowDefinitions="Auto,Auto,*,Auto">
