+++ Source/LibationWinForms/Properties/PublishProfiles/FolderProfile.pubxml
@@ -5,7 +5,7 @@ https://go.microsoft.com/fwlink/?LinkID=208121.
 <Project>
   <PropertyGroup>
     <Configuration>Release</Configuration>
-    <Platform>x64</Platform>
+    <Platform>Any CPU</Platform>
     <PublishDir>..\bin\publish\</PublishDir>
     <PublishProtocol>FileSystem</PublishProtocol>
     <TargetFramework>net6.0-windows</TargetFramework>
